Refactor to introduce a logger
