+++ lib/run_loop/core.rb
@@ -8,12 +8,6 @@ require 'ap'
 
 module RunLoop
 
-  class TimeoutError < RuntimeError
-  end
-
-  class WriteFailedError < RuntimeError
-  end
-
   module Core
 
     START_DELIMITER = "OUTPUT_JSON:\n"
@@ -49,7 +43,9 @@ module RunLoop
       # RunLoop::Version overrides '=='
       options_to_log[:xcode] = xctools.xcode_version.to_s
       options_to_log[:xcode_path] = xctools.xcode_developer_dir
-      ap(options_to_log, {:sort_keys => true})
+      message = options_to_log.ai({:sort_keys => true})
+      logger = options[:logger]
+      RunLoop::Logging.log_debug(logger, "\n" + message)
     end
 
     # @deprecated since 1.0.0
@@ -104,6 +100,7 @@ module RunLoop
     # @raise [RunLoop::IncompatibleArchitecture] Raises an error if the
     #  application binary is not compatible with the target simulator.
     def self.expect_compatible_simulator_architecture(launch_options, sim_control)
+      logger = launch_options[:logger]
       if sim_control.xcode_version_gte_6?
         sim_identifier = launch_options[:udid]
         simulator = sim_control.simulators.find do |simulator|
@@ -117,14 +114,10 @@ module RunLoop
 
         lipo = RunLoop::Lipo.new(launch_options[:bundle_dir_or_bundle_id])
         lipo.expect_compatible_arch(simulator)
-        if RunLoop::Environment.debug?
-          puts "Simulator instruction set '#{simulator.instruction_set}' is compatible with #{lipo.info}"
-        end
+        RunLoop::Logging.log_debug(logger, "Simulator instruction set '#{simulator.instruction_set}' is compatible with #{lipo.info}")
         true
       else
-        if RunLoop::Environment.debug?
-          puts "Xcode #{sim_control.xctools.xcode_version} detected; skipping simulator architecture check."
-        end
+        RunLoop::Logging.log_debug(logger, "Xcode #{sim_control.xctools.xcode_version} detected; skipping simulator architecture check.")
         false
       end
     end
@@ -132,6 +125,7 @@ module RunLoop
     def self.run_with_options(options)
       before = Time.now
 
+      logger = options[:logger]
       sim_control ||= options[:sim_control] || RunLoop::SimControl.new
       xctools ||= options[:xctools] || sim_control.xctools
 
@@ -170,7 +164,7 @@ module RunLoop
       uia_strategy = options[:uia_strategy]
       if uia_strategy == :host
         create_uia_pipe(repl_path)
-        RunLoop::HostCache.default.clear
+        RunLoop::HostCache.default.clear unless RunLoop::Environment.xtc?
       end
 
       cal_script = File.join(SCRIPTS_PATH, 'calabash_script_uia.js')
@@ -189,9 +183,7 @@ module RunLoop
       log_file ||= File.join(results_dir, 'run_loop.out')
 
       after = Time.now
-      if RunLoop::Environment.debug?
-        puts "Preparation took #{after-before} seconds"
-      end
+      RunLoop::Logger.log_debug(logger, "Preparation took #{after-before} seconds")
 
       discovered_options =
             {
@@ -215,7 +207,7 @@ module RunLoop
 
       automation_template = automation_template(xctools)
 
-      log_header("Starting on #{device_target} App: #{bundle_dir_or_bundle_id}")
+      RunLoop::Logging.log_header(logger, "Starting on #{device_target} App: #{bundle_dir_or_bundle_id}")
 
       pid = instruments.spawn(automation_template, merged_options, log_file)
 
@@ -247,17 +239,12 @@ module RunLoop
           end
         end
       rescue TimeoutError => e
-        if RunLoop::Environment.debug?
-          puts "Failed to launch."
-          puts "#{e}: #{e && e.message}"
-        end
+        RunLoop::Logging.log_debug(logger, "Failed to launch. #{e}: #{e && e.message}")
         raise TimeoutError, "Time out waiting for UIAutomation run-loop to Start. \n Logfile #{log_file} \n\n #{File.read(log_file)}\n"
       end
 
-      if RunLoop::Environment.debug?
-        puts "Launching took #{Time.now-before} seconds"
-      end
-
+      RunLoop::Logging.log_debug(logger, "Launching took #{Time.now-before} seconds")
+      
       dylib_path = self.dylib_path_from_options(merged_options)
 
       if dylib_path
@@ -437,21 +424,20 @@ module RunLoop
       repl_path = run_loop[:repl_path]
       index = run_loop[:index]
       cmd_str = "#{index}:#{escape_host_command(cmd)}"
-      should_log = RunLoop::Environment.debug?
-      RunLoop.log_info(logger, cmd_str) if should_log
+      RunLoop::Logging.log_debug(logger, cmd_str)
       write_succeeded = false
       2.times do |i|
-        RunLoop.log_info(logger, "Trying write of command #{cmd_str} at index #{index}") if should_log
+        RunLoop::Logging.log_debug(logger, "Trying write of command #{cmd_str} at index #{index}")
         File.open(repl_path, 'w') { |f| f.puts(cmd_str) }
         write_succeeded = validate_index_written(run_loop, index, logger)
         break if write_succeeded
       end
       unless write_succeeded
-        RunLoop.log_info(logger, 'Failing...Raising RunLoop::WriteFailedError') if should_log
+        RunLoop::Logging.log_debug(logger, 'Failing...Raising RunLoop::WriteFailedError')
         raise RunLoop::WriteFailedError.new("Trying write of command #{cmd_str} at index #{index}")
       end
       run_loop[:index] = index + 1
-      RunLoop::HostCache.default.write(run_loop)
+      RunLoop::HostCache.default.write(run_loop) unless RunLoop::Environment.xtc?
       index
     end
 
@@ -460,10 +446,10 @@ module RunLoop
         Timeout::timeout(10, TimeoutError) do
           Core.read_response(run_loop, index, 10, 'last_index')
         end
-        RunLoop.log_info(logger, "validate index written for index #{index} ok")
+        RunLoop::Logging.log_debug(logger, "validate index written for index #{index} ok")
         return true
       rescue TimeoutError => _
-        RunLoop.log_info(logger, "validate index written for index #{index} failed. Retrying.")
+        RunLoop::Logging.log_debug(logger, "validate index written for index #{index} failed. Retrying.")
         return false
       end
     end
@@ -485,9 +471,7 @@ module RunLoop
           next
         end
 
-
         size = File.size(log_file)
-
         output = File.read(log_file, size-offset, offset)
 
         if /AXError: Could not auto-register for pid status change/.match(output)
@@ -545,7 +529,7 @@ module RunLoop
       end
 
       run_loop[:initial_offset] = offset
-      RunLoop::HostCache.default.write(run_loop)
+      RunLoop::HostCache.default.write(run_loop) unless RunLoop::Environment.xtc?
       result
     end
 
@@ -588,20 +572,6 @@ module RunLoop
       raise msgs.join("\n")
     end
 
-    def self.log(message)
-      if RunLoop::Environment.debug?
-        puts "#{Time.now } #{message}"
-        $stdout.flush
-      end
-    end
-
-    def self.log_header(message)
-      if RunLoop::Environment.debug?
-        puts "\n\e[#{35}m### #{message} ###\e[0m"
-        $stdout.flush
-      end
-    end
-
     # @deprecated 1.0.5
     def self.ensure_instruments_not_running!
       RunLoop::Instruments.new.kill_instruments
@@ -617,170 +587,4 @@ module RunLoop
     end
   end
 
-  def self.default_script_for_uia_strategy(uia_strategy)
-    case uia_strategy
-      when :preferences
-        Core.script_for_key(:run_loop_fast_uia)
-      when :host
-        Core.script_for_key(:run_loop_host)
-      when :shared_element
-        Core.script_for_key(:run_loop_shared_element)
-      else
-        Core.script_for_key(:run_loop_basic)
-    end
-  end
-
-  def self.run(options={})
-
-    if RunLoop::Instruments.new.instruments_app_running?
-      msg =
-            [
-                  "Please quit the Instruments.app.",
-                  "If Instruments.app is open, the instruments command line",
-                  "tool cannot take control of your application."
-            ]
-      raise msg.join("\n")
-    end
-
-    uia_strategy = options[:uia_strategy]
-    if options[:script]
-      script = validate_script(options[:script])
-    else
-      if uia_strategy
-        script = default_script_for_uia_strategy(uia_strategy)
-      else
-        if options[:calabash_lite]
-          uia_strategy = :host
-          script = Core.script_for_key(:run_loop_host)
-        else
-          uia_strategy = :preferences
-          script = default_script_for_uia_strategy(uia_strategy)
-        end
-      end
-    end
-    # At this point, 'script' has been chosen, but uia_strategy might not
-    unless uia_strategy
-      desired_script = options[:script]
-      if desired_script.is_a?(String) #custom path to script
-        uia_strategy = :host
-      elsif desired_script == :run_loop_host
-        uia_strategy = :host
-      elsif desired_script == :run_loop_fast_uia
-        uia_strategy = :preferences
-      elsif desired_script == :run_loop_shared_element
-        uia_strategy = :shared_element
-      else
-        raise "Inconsistent state: desired script #{desired_script} has not uia_strategy"
-      end
-    end
-    # At this point script and uia_strategy selected
-
-    options[:script] = script
-    options[:uia_strategy] = uia_strategy
-
-    Core.run_with_options(options)
-  end
-
-  def self.send_command(run_loop, cmd, options={timeout: 60}, num_retries=0, last_error=nil)
-    if num_retries > 3
-      if last_error
-        raise last_error
-      else
-        raise "Max retries exceeded #{num_retries} > 3. No error recorded."
-      end
-    end
-
-    if options.is_a?(Numeric)
-      options = {timeout: options}
-    end
-
-    if not cmd.is_a?(String)
-      raise "Illegal command #{cmd} (must be a string)"
-    end
-
-    if not options.is_a?(Hash)
-      raise "Illegal options #{options} (must be a Hash (or number for compatibility))"
-    end
-
-    timeout = options[:timeout] || 60
-    logger = options[:logger]
-    interrupt_retry_timeout = options[:interrupt_retry_timeout] || 25
-
-    expected_index = run_loop[:index]
-    result = nil
-    begin
-      expected_index = Core.write_request(run_loop, cmd, logger)
-    rescue RunLoop::WriteFailedError, Errno::EINTR => write_error
-      # Attempt recover from interrupt by attempting to read result (assuming write went OK)
-      # or retry if attempted read result fails
-      run_loop[:index] = expected_index # restore expected index in case it changed
-      log_info(logger, "Core.write_request failed: #{write_error}. Attempting recovery...")
-      log_info(logger, "Attempting read in case the request was received... Please wait (#{interrupt_retry_timeout})...")
-      begin
-        Timeout::timeout(interrupt_retry_timeout, TimeoutError) do
-          result = Core.read_response(run_loop, expected_index)
-        end
-        # Update run_loop expected index since we succeeded in reading the index
-        run_loop[:index] = expected_index + 1
-        log_info(logger, "Did read response for interrupted request of index #{expected_index}... Proceeding.")
-        return result
-      rescue TimeoutError => _
-        log_info(logger, "Read did not result in a response for index #{expected_index}... Retrying send_command...")
-        return send_command(run_loop, cmd, options, num_retries+1, write_error)
-      end
-    end
-
-
-    begin
-      Timeout::timeout(timeout, TimeoutError) do
-        result = Core.read_response(run_loop, expected_index)
-      end
-    rescue TimeoutError => _
-      raise TimeoutError, "Time out waiting for UIAutomation run-loop for command #{cmd}. Waiting for index:#{expected_index}"
-    end
-
-    result
-  end
-
-  def self.stop(run_loop, out=Dir.pwd)
-    return if run_loop.nil?
-    results_dir = run_loop[:results_dir]
-    dest = out
-
-    RunLoop::Instruments.new.kill_instruments
-
-    FileUtils.mkdir_p(dest)
-    if results_dir
-      pngs = Dir.glob(File.join(results_dir, 'Run 1', '*.png'))
-    else
-      pngs = []
-    end
-    FileUtils.cp(pngs, dest) if pngs and pngs.length > 0
-  end
-
-
-  def self.validate_script(script)
-    if script.is_a?(String)
-      unless File.exist?(script)
-        raise "Unable to find file: #{script}"
-      end
-    elsif script.is_a?(Symbol)
-      script = Core.script_for_key(script)
-      unless script
-        raise "Unknown script for symbol: #{script}. Options: #{Core::SCRIPTS.keys.join(', ')}"
-      end
-    else
-      raise "Script must be a symbol or path: #{script}"
-    end
-    script
-  end
-
-  def self.log_info(logger, message)
-    msg = "#{Time.now}: #{message}"
-    if logger && logger.respond_to?(:info)
-      logger.info(msg)
-    else
-      puts msg if RunLoop::Environment.debug?
-    end
-  end
 end
